projects/182-healthcare-coi-analytics-report-template/requirements.txt
# Healthcare COI Analytics Requirements
# Python 3.8+

# Core data processing
pandas>=1.5.0
numpy>=1.23.0

# Google Cloud
google-cloud-bigquery>=3.0.0
google-cloud-storage>=2.0.0

# Configuration and environment
python-dotenv>=0.19.0
pyyaml>=6.0

# Excel processing
openpyxl>=3.0.0
xlsxwriter>=3.0.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.12.0
plotly>=5.0.0

# PDF generation
reportlab>=3.6.0
markdown>=3.4.0
weasyprint>=57.0  # Optional: for HTML to PDF

# Statistical analysis
scipy>=1.9.0
statsmodels>=0.13.0

# Utilities
tqdm>=4.64.0  # Progress bars
tabulate>=0.9.0  # Table formatting

# Development (optional)
ipython>=8.0.0
jupyter>=1.0.0
black>=22.0.0  # Code formatting
pytest>=7.0.0  # Testing